Add the base BudouX Web component
